+++ docs/en/docs/release-notes.md
@@ -11,7 +11,7 @@ hide:
 
 - Support for multiple cookie headers in the `Request` object. This follows the [RFC 7540](https://datatracker.ietf.org/doc/html/rfc7540).
 
-## Changed
+### Changed
 
 - Core implementation for EventStreamResponse to support multiple headers and parameters.